+++ lib/fluent/plugin/out_influxdb.rb
@@ -1,55 +1,65 @@
 # encoding: UTF-8
 require 'date'
 require 'influxdb'
-require 'fluent/output'
+require 'fluent/plugin/output'
 require 'fluent/mixin'
 
-class Fluent::InfluxdbOutput < Fluent::BufferedOutput
+class Fluent::Plugin::InfluxdbOutput < Fluent::Plugin::Output
   Fluent::Plugin.register_output('influxdb', self)
 
-  include Fluent::HandleTagNameMixin
+  helpers :compat_parameters
 
-  config_param :host, :string,  :default => 'localhost',
-               :desc => "The IP or domain of influxDB, separate with comma."
-  config_param :port, :integer,  :default => 8086,
-               :desc => "The HTTP port of influxDB."
-  config_param :dbname, :string,  :default => 'fluentd',
-               :desc => <<-DESC
+  DEFAULT_BUFFER_TYPE = "memory"
+
+  config_param :host, :string,  default: 'localhost',
+               desc: "The IP or domain of influxDB, separate with comma."
+  config_param :port, :integer,  default: 8086,
+               desc: "The HTTP port of influxDB."
+  config_param :dbname, :string,  default: 'fluentd',
+               desc: <<-DESC
 The database name of influxDB.
 You should create the database and grant permissions at first.
 DESC
-  config_param :user, :string,  :default => 'root',
-               :desc => "The DB user of influxDB, should be created manually."
-  config_param :password, :string,  :default => 'root', :secret => true,
-               :desc => "The password of the user."
-  config_param :retry, :integer, :default => nil,
-               :desc => 'The finite number of retry times. default is infinite'
-  config_param :time_key, :string, :default => 'time',
-               :desc => 'Use value of this tag if it exists in event instead of event timestamp'
-  config_param :time_precision, :string, :default => 's',
-               :desc => <<-DESC
+  config_param :measurement, :string, default: nil,
+               desc: "The measurement name to insert events. If not specified, fluentd's tag is used"
+  config_param :user, :string,  default: 'root',
+               desc: "The DB user of influxDB, should be created manually."
+  config_param :password, :string,  default: 'root', secret: true,
+               desc: "The password of the user."
+  config_param :retry, :integer, default: nil,
+               desc: 'The finite number of retry times. default is infinite'
+  config_param :time_key, :string, default: 'time',
+               desc: 'Use value of this tag if it exists in event instead of event timestamp'
+  config_param :time_precision, :string, default: 's',
+               desc: <<-DESC
 The time precision of timestamp.
 You should specify either hour (h), minutes (m), second (s),
-millisecond (ms), microsecond (u), or nanosecond (n).
+millisecond (ms), microsecond (u), or nanosecond (ns).
 DESC
-  config_param :use_ssl, :bool, :default => false,
-               :desc => "Use SSL when connecting to influxDB."
-  config_param :verify_ssl, :bool, :default => true,
-               :desc => "Enable/Disable SSL Certs verification when connecting to influxDB via SSL."
+  config_param :use_ssl, :bool, default: false,
+               desc: "Use SSL when connecting to influxDB."
+  config_param :verify_ssl, :bool, default: true,
+               desc: "Enable/Disable SSL Certs verification when connecting to influxDB via SSL."
   config_param :keys_mapping, :hash, :default => {},
-               :desc => "The names of the keys mapping to map existing tags to those from influxDB."
-  config_param :tag_keys, :array, :default => [],
-               :desc => "The names of the keys to use as influxDB tags."
-  config_param :sequence_tag, :string, :default => nil,
-               :desc => <<-DESC
+               desc: "The names of the keys mapping to map existing tags to those from influxDB."
+  config_param :auto_tags, :bool, default: false,
+               desc: "Enable/Disable auto-tagging behaviour which makes strings tags."
+  config_param :tag_keys, :array, default: [],
+               desc: "The names of the keys to use as influxDB tags."
+  config_param :sequence_tag, :string, default: nil,
+               desc: <<-DESC
 The name of the tag whose value is incremented for the consecutive simultaneous
 events and reset to zero for a new event with the different timestamp.
 DESC
-  config_param :retention_policy_key, :string, :default => nil,
-               :desc => "The key of the key in the record that stores the retention policy name"
-  config_param :default_retention_policy, :string, :default => nil,
-               :desc => "The name of the default retention policy"
+  config_param :retention_policy_key, :string, default: nil,
+               desc: "The key of the key in the record that stores the retention policy name"
+  config_param :default_retention_policy, :string, default: nil,
+               desc: "The name of the default retention policy"
 
+  config_section :buffer do
+    config_set_default :@type, DEFAULT_BUFFER_TYPE
+    config_set_default :chunk_keys, ['tag']
+  end
 
   def initialize
     super
@@ -58,7 +68,10 @@ DESC
   end
 
   def configure(conf)
+    compat_parameters_convert(conf, :buffer)
     super
+    @time_precise = time_precise_lambda()
+    raise Fluent::ConfigError, "'tag' in chunk_keys is required." if not @chunk_key_tag
   end
 
   def start
@@ -82,7 +95,7 @@ DESC
       unless existing_databases.include? @dbname
         raise Fluent::ConfigError, 'Database ' + @dbname + ' doesn\'t exist. Create it first, please. Existing databases: ' + existing_databases.join(',')
       end
-    rescue InfluxDB::AuthenticationError
+    rescue InfluxDB::AuthenticationError, InfluxDB::Error
       $log.info "skip database presence check because '#{@user}' user doesn't have admin privilege. Check '#{@dbname}' exists on influxdb"
     end
   end
@@ -94,7 +107,7 @@ DESC
     if record.empty? || record.has_value?(nil)
       FORMATTED_RESULT_FOR_INVALID_RECORD
     else
-      [tag, time, record].to_msgpack
+      [precision_time(time), record].to_msgpack
     end
   end
 
@@ -103,11 +116,20 @@ DESC
     @influxdb.stop!
   end
 
+  def formatted_to_msgpack_binary
+    true
+  end
+
+  def multi_workers_ready?
+    true
+  end
+
   def write(chunk)
     points = []
-    chunk.msgpack_each do |tag, time, record|
+    tag = chunk.metadata.tag
+    chunk.msgpack_each do |time, record|
       timestamp = record.delete(@time_key) || time
-      if @tag_keys.empty? and @keys_mapping.empty?
+      if @tag_keys.empty? && !@auto_tags && @keys_mapping.empty?
         values = record
         tags = {}
       else
@@ -115,7 +137,7 @@ DESC
         tags = {}
         record.each_pair do |k, v|
           k = @keys_mapping.fetch(k, k)
-          if @tag_keys.include?(k)
+          if (@auto_tags && v.is_a?(String)) || @tag_keys.include?(k)
             # If the tag value is not nil, empty, or a space, add the tag
             if v.to_s.strip != ''
               tags[k] = v
@@ -134,11 +156,17 @@ DESC
         tags[@sequence_tag] = @seq
         @prev_timestamp = timestamp
       end
+
+      if values.empty?
+          log.warn "Skip record '#{record}', because InfluxDB requires at least one value in raw"
+          next
+      end
+
       point = {
-        :timestamp => timestamp,
-        :series    => tag,
-        :values    => values,
-        :tags      => tags,
+        timestamp: timestamp,
+        series: @measurement || tag,
+        values: values,
+        tags: tags,
       }
       retention_policy = @default_retention_policy
       unless @retention_policy_key.nil?
@@ -165,5 +193,31 @@ DESC
         @influxdb.write_points(points, nil, @default_retention_policy)
       end
     end
+  end
+
+  def time_precise_lambda()
+    case @time_precision.to_sym
+    when :h then
+      lambda{|nstime| nstime / (10 ** 9) / (60 ** 2) }
+    when :m then
+      lambda{|nstime| nstime / (10 ** 9) / 60 }
+    when :s then
+      lambda{|nstime| nstime / (10 ** 9) }
+    when :ms then
+      lambda{|nstime| nstime / (10 ** 6) }
+    when :u then
+      lambda{|nstime| nstime / (10 ** 3) }
+    when :ns then
+      lambda{|nstime| nstime }
+    else
+      raise Fluent::ConfigError, 'time_precision ' + @time_precision + ' is invalid.' +
+        'should specify either either hour (h), minutes (m), second (s), millisecond (ms), microsecond (u), or nanosecond (ns)'
+    end
+  end
+
+  def precision_time(time)
+    # nsec is supported from v0.14
+    nstime = time * (10 ** 9) + (time.is_a?(Integer) ? 0 : time.nsec)
+    @time_precise.call(nstime)
   end
 end