+++ test/plugin/test_out_influxdb.rb
@@ -20,6 +20,26 @@ class InfluxdbOutputTest < Test::Unit::TestCase
     end
   end
 
+  class DummyInfluxdbOutput < Fluent::Plugin::InfluxdbOutput
+    attr_reader :influxdb
+    def configure(conf)
+      @influxdb = DummyInfluxDBClient.new()
+      super
+    end
+
+    def format(tag, time, record)
+      super
+    end
+
+    def formatted_to_msgpack_binary
+      true
+    end
+
+    def write(chunk)
+      super
+    end
+  end
+
   def setup
     Fluent::Test.setup
   end
@@ -35,18 +55,12 @@ class InfluxdbOutputTest < Test::Unit::TestCase
     time_precision s
   ]
 
-  def create_raw_driver(conf=CONFIG, tag='test')
-    Fluent::Test::BufferedOutputTestDriver.new(Fluent::InfluxdbOutput, tag).configure(conf)
+  def create_raw_driver(conf=CONFIG)
+    Fluent::Test::Driver::Output.new(Fluent::Plugin::InfluxdbOutput).configure(conf)
   end
 
-  def create_driver(conf=CONFIG, tag='test')
-    Fluent::Test::BufferedOutputTestDriver.new(Fluent::InfluxdbOutput, tag) do
-      attr_reader :influxdb
-      def configure(conf)
-        @influxdb = DummyInfluxDBClient.new()
-        super
-      end
-    end.configure(conf)
+  def create_driver(conf=CONFIG)
+    Fluent::Test::Driver::Output.new(DummyInfluxdbOutput).configure(conf)
   end
 
   def test_configure
@@ -60,40 +74,91 @@ class InfluxdbOutputTest < Test::Unit::TestCase
     assert_equal('xxxxxx', driver.instance.config['password'])
   end
 
-  def test_format
-    driver = create_driver(CONFIG, 'test')
-    time = Time.parse('2011-01-02 13:14:15 UTC').to_i
-
-    driver.emit({'a' => 1}, time)
-    driver.emit({'a' => 2}, time)
-
-    driver.expect_format(['test', time, {'a' => 1}].to_msgpack)
-    driver.expect_format(['test', time, {'a' => 2}].to_msgpack)
-
-    driver.run
+  def test_configure_without_tag_chunk_key
+    assert_raise(Fluent::ConfigError) do
+      create_raw_driver %[
+        dbname test
+        user  testuser
+        password  mypwd
+        <buffer arbitrary_key>
+          @type memory
+        </buffer>
+      ]
+    end
   end
 
-  def test_write
-    driver = create_driver(CONFIG, 'input.influxdb')
+  def test_format
+    driver = create_driver(CONFIG)
+    time = event_time('2011-01-02 13:14:15 UTC')
 
-    time = Time.parse("2011-01-02 13:14:15 UTC").to_i
-    driver.emit({'a' => 1}, time)
-    driver.emit({'a' => 2}, time)
+    driver.run(default_tag: 'test') do
+      driver.feed(time, {'a' => 1})
+      driver.feed(time, {'a' => 2})
+    end
 
-    data = driver.run
+    assert_equal [[time, {'a' => 1}].to_msgpack,
+                  [time, {'a' => 2}].to_msgpack], driver.formatted
+  end
+
+  sub_test_case "#write" do
+    test "buffer" do
+      driver = create_driver(CONFIG)
+
+      time = event_time("2011-01-02 13:14:15 UTC")
+      driver.run(default_tag: 'input.influxdb') do
+        driver.feed(time, {'a' => 1})
+        driver.feed(time, {'a' => 2})
+      end
+
+      assert_equal([
+        [
+          [
+            {
+              timestamp: time,
+              series: 'input.influxdb',
+              tags: {},
+              values: {'a' => 1}
+            },
+            {
+              timestamp: time,
+              series: 'input.influxdb',
+              tags: {},
+              values: {'a' => 2}
+            },
+          ],
+          nil,
+          nil
+        ]
+      ], driver.instance.influxdb.points)
+    end
+  end
+
+  def test_write_with_measurement
+    config_with_measurement = %Q(
+      #{CONFIG}
+      measurement test
+    )
+
+    driver = create_driver(config_with_measurement)
+
+    time = event_time('2011-01-02 13:14:15 UTC')
+    driver.run(default_tag: 'input.influxdb') do
+      driver.feed(time, {'a' => 1})
+      driver.feed(time, {'a' => 2})
+    end
 
     assert_equal([
       [
         [
           {
             :timestamp => time,
-            :series    => 'input.influxdb',
+            :series    => 'test',
             :tags      => {},
             :values    => {'a' => 1}
           },
           {
             :timestamp => time,
-            :series    => 'input.influxdb',
+            :series    => 'test',
             :tags      => {},
             :values    => {'a' => 2}
           },
@@ -102,24 +167,66 @@ class InfluxdbOutputTest < Test::Unit::TestCase
         nil
       ]
     ], driver.instance.influxdb.points)
-
   end
 
   def test_empty_tag_keys
     config_with_tags = %Q(
       #{CONFIG}
-
       tag_keys ["b"]
     )
 
-    driver = create_driver(config_with_tags, 'input.influxdb')
+    driver = create_driver(config_with_tags)
 
-    time = Time.parse("2011-01-02 13:14:15 UTC").to_i
-    driver.emit({'a' => 1, 'b' => ''}, time)
-    driver.emit({'a' => 2, 'b' => 1}, time)
-    driver.emit({'a' => 3, 'b' => ' '}, time)
+    time = event_time("2011-01-02 13:14:15 UTC")
+    driver.run(default_tag: 'input.influxdb') do
+      driver.feed(time, {'a' => 1, 'b' => ''})
+      driver.feed(time, {'a' => 2, 'b' => 1})
+      driver.feed(time, {'a' => 3, 'b' => ' '})
+    end
 
-    data = driver.run
+    assert_equal([
+      [
+        [
+          {
+            timestamp: time,
+            series: 'input.influxdb',
+            values: {'a' => 1},
+            tags: {},
+          },
+          {
+            timestamp: time,
+            series: 'input.influxdb',
+            values: {'a' => 2},
+            tags: {'b' => 1},
+          },
+          {
+            timestamp: time,
+            series: 'input.influxdb',
+            values: {'a' => 3},
+            tags: {},
+          },
+        ],
+        nil,
+        nil
+      ]
+    ], driver.instance.influxdb.points)
+  end
+
+  def test_auto_tagging
+    config_with_tags = %Q(
+      #{CONFIG}
+
+      auto_tags true
+    )
+
+    driver = create_driver(config_with_tags)
+
+    time = event_time("2011-01-02 13:14:15 UTC")
+    driver.run(default_tag: 'input.influxdb') do
+      driver.feed(time, {'a' => 1, 'b' => '1'})
+      driver.feed(time, {'a' => 2, 'b' => 1})
+      driver.feed(time, {'a' => 3, 'b' => ' '})
+    end
 
     assert_equal([
       [
@@ -127,14 +234,15 @@ class InfluxdbOutputTest < Test::Unit::TestCase
           {
             :timestamp => time,
             :series    => 'input.influxdb',
+
             :values    => {'a' => 1},
-            :tags      => {},
+            :tags      => {'b' => '1'},
           },
           {
             :timestamp => time,
             :series    => 'input.influxdb',
-            :values    => {'a' => 2},
-            :tags      => {'b' => 1},
+            :values    => {'a' => 2, 'b' => 1},
+            :tags      => {},
           },
           {
             :timestamp => time,
@@ -142,6 +250,39 @@ class InfluxdbOutputTest < Test::Unit::TestCase
             :values    => {'a' => 3},
             :tags      => {},
           },
+
+        ],
+        nil,
+        nil
+      ]
+    ], driver.instance.influxdb.points)
+  end
+
+  def test_ignore_empty_values
+    config_with_tags = %Q(
+      #{CONFIG}
+
+      tag_keys ["b"]
+    )
+
+    driver = create_driver(config_with_tags)
+
+    time = event_time('2011-01-02 13:14:15 UTC')
+    driver.run(default_tag: 'input.influxdb') do
+      driver.feed(time, {'b' => '3'})
+      driver.feed(time, {'a' => 2, 'b' => 1})
+    end
+
+    assert_equal([
+      [
+        [
+          {
+            :timestamp => time,
+            :series    => 'input.influxdb',
+
+            :values    => {'a' => 2},
+            :tags      => {'b' => 1},
+          }
         ],
         nil,
         nil
@@ -158,46 +299,47 @@ class InfluxdbOutputTest < Test::Unit::TestCase
       user  testuser
       password  mypwd
       use_ssl false
-      time_precision s 
+      time_precision s
       sequence_tag _seq
     ]
-    driver = create_driver(config, 'input.influxdb')
+    driver = create_driver(config)
 
-    time = Time.parse("2011-01-02 13:14:15 UTC").to_i
-    driver.emit({'a' => 1}, time)
-    driver.emit({'a' => 2}, time)
+    time = event_time("2011-01-02 13:14:15 UTC")
+    next_time = Fluent::EventTime.new(time + 1)
+    driver.run(default_tag: 'input.influxdb') do
+      driver.feed(time, {'a' => 1})
+      driver.feed(time, {'a' => 2})
 
-    driver.emit({'a' => 1}, time + 1)
-    driver.emit({'a' => 2}, time + 1)
-
-    data = driver.run
+      driver.feed(next_time, {'a' => 1})
+      driver.feed(next_time, {'a' => 2})
+    end
 
     assert_equal([
       [
         [
           {
-            :timestamp => time,
-            :series    => 'input.influxdb',
-            :values    => {'a' => 1},
-            :tags      => {'_seq' => 0},
+            timestamp: time,
+            series: 'input.influxdb',
+            values: {'a' => 1},
+            tags: {'_seq' => 0},
           },
           {
-            :timestamp => time,
-            :series    => 'input.influxdb',
-            :values    => {'a' => 2},
-            :tags      => {'_seq' => 1},
+            timestamp: time,
+            series: 'input.influxdb',
+            values: {'a' => 2},
+            tags: {'_seq' => 1},
           },
           {
-            :timestamp => time + 1,
-            :series    => 'input.influxdb',
-            :values    => {'a' => 1},
-            :tags      => {'_seq' => 0},
+            timestamp: time + 1,
+            series: 'input.influxdb',
+            values: {'a' => 1},
+            tags: {'_seq' => 0},
           },
           {
-            :timestamp => time + 1,
-            :series    => 'input.influxdb',
-            :values    => {'a' => 2},
-            :tags      => {'_seq' => 1},
+            timestamp: time + 1,
+            series: 'input.influxdb',
+            values: {'a' => 2},
+            tags: {'_seq' => 1},
           }
         ],
         nil,
@@ -211,28 +353,28 @@ class InfluxdbOutputTest < Test::Unit::TestCase
     config = CONFIG + "\n" + %[
       default_retention_policy ephemeral_1d
     ]
-    driver = create_driver(config, 'input.influxdb')
+    driver = create_driver(config)
 
-    time = Time.parse("2011-01-02 13:14:15 UTC").to_i
-    driver.emit({'a' => 1}, time)
-    driver.emit({'a' => 2}, time)
-
-    data = driver.run
+    time = event_time("2011-01-02 13:14:15 UTC")
+    driver.run(default_tag: 'input.influxdb') do
+      driver.feed(time, {'a' => 1})
+      driver.feed(time, {'a' => 2})
+    end
 
     assert_equal([
       [
         [
           {
-            :timestamp => time,
-            :series    => 'input.influxdb',
-            :tags      => {},
-            :values    => {'a' => 1}
+            timestamp: time,
+            series: 'input.influxdb',
+            tags: {},
+            values: {'a' => 1}
           },
           {
-            :timestamp => time,
-            :series    => 'input.influxdb',
-            :tags      => {},
-            :values    => {'a' => 2}
+            timestamp: time,
+            series: 'input.influxdb',
+            tags: {},
+            values: {'a' => 2}
           },
         ],
         nil,
@@ -246,23 +388,23 @@ class InfluxdbOutputTest < Test::Unit::TestCase
     config = CONFIG + "\n" + %[
       retention_policy_key rp
     ]
-    driver = create_driver(config, 'input.influxdb')
+    driver = create_driver(config)
 
-    time = Time.parse("2011-01-02 13:14:15 UTC").to_i
-    driver.emit({'a' => 1}, time)
-    driver.emit({'a' => 2, 'rp' => 'ephemeral_1d'}, time)
-    driver.emit({'a' => 3, 'rp' => 'ephemeral_1m'}, time)
-
-    data = driver.run
+    time = event_time("2011-01-02 13:14:15 UTC")
+    driver.run(default_tag: 'input.influxdb') do
+      driver.feed(time, {'a' => 1})
+      driver.feed(time, {'a' => 2, 'rp' => 'ephemeral_1d'})
+      driver.feed(time, {'a' => 3, 'rp' => 'ephemeral_1m'})
+    end
 
     assert_equal([
       [
         [
           {
-            :timestamp => time,
-            :series    => 'input.influxdb',
-            :tags      => {},
-            :values    => {'a' => 1},
+            timestamp: time,
+            series: 'input.influxdb',
+            tags: {},
+            values: {'a' => 1},
           }
         ],
         nil,
@@ -271,10 +413,10 @@ class InfluxdbOutputTest < Test::Unit::TestCase
       [
         [
           {
-            :timestamp => time,
-            :series    => 'input.influxdb',
-            :tags      => {},
-            :values    => {'a' => 2},
+            timestamp: time,
+            series: 'input.influxdb',
+            tags: {},
+            values: {'a' => 2},
           }
         ],
         nil,
@@ -283,10 +425,10 @@ class InfluxdbOutputTest < Test::Unit::TestCase
       [
         [
           {
-            :timestamp => time,
-            :series    => 'input.influxdb',
-            :tags      => {},
-            :values    => {'a' => 3},
+            timestamp: time,
+            series: 'input.influxdb',
+            tags: {},
+            values: {'a' => 3},
           }
         ],
         nil,
@@ -301,30 +443,30 @@ class InfluxdbOutputTest < Test::Unit::TestCase
       default_retention_policy ephemeral_1d
       retention_policy_key rp
     ]
-    driver = create_driver(config, 'input.influxdb')
+    driver = create_driver(config)
 
-    time = Time.parse("2011-01-02 13:14:15 UTC").to_i
-    driver.emit({'a' => 1}, time)
-    driver.emit({'a' => 2, 'rp' => 'ephemeral_1d'}, time)
-    driver.emit({'a' => 3, 'rp' => 'ephemeral_1m'}, time)
-    driver.emit({'a' => 4}, time)
-
-    data = driver.run
+    time = event_time("2011-01-02 13:14:15 UTC")
+    driver.run(default_tag: 'input.influxdb') do
+      driver.feed(time, {'a' => 1})
+      driver.feed(time, {'a' => 2, 'rp' => 'ephemeral_1d'})
+      driver.feed(time, {'a' => 3, 'rp' => 'ephemeral_1m'})
+      driver.feed(time, {'a' => 4})
+    end
 
     assert_equal([
       [
         [
           {
-            :timestamp => time,
-            :series    => 'input.influxdb',
-            :tags      => {},
-            :values    => {'a' => 1},
+            timestamp: time,
+            series: 'input.influxdb',
+            tags: {},
+            values: {'a' => 1},
           },
           {
-            :timestamp => time,
-            :series    => 'input.influxdb',
-            :tags      => {},
-            :values    => {'a' => 2},
+            timestamp: time,
+            series: 'input.influxdb',
+            tags: {},
+            values: {'a' => 2},
           }
         ],
         nil,
@@ -333,10 +475,10 @@ class InfluxdbOutputTest < Test::Unit::TestCase
       [
         [
           {
-            :timestamp => time,
-            :series    => 'input.influxdb',
-            :tags      => {},
-            :values    => {'a' => 3},
+            timestamp: time,
+            series: 'input.influxdb',
+            tags: {},
+            values: {'a' => 3},
           }
         ],
         nil,
@@ -345,10 +487,10 @@ class InfluxdbOutputTest < Test::Unit::TestCase
       [
         [
           {
-            :timestamp => time,
-            :series    => 'input.influxdb',
-            :tags      => {},
-            :values    => {'a' => 4},
+            timestamp: time,
+            series: 'input.influxdb',
+            tags: {},
+            values: {'a' => 4},
           }
         ],
         nil,